Move helpers to HTML Pipelines
